# backend/api-spec.yaml
openapi: 3.0.3
info:
  title: TuneRecommend
  version: 1.0.0
paths:
...
  /api/song/{id}:
    get:
      summary: Get song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song object
          content:
            application/json:
              example:
                id: 1
                mbid: "song-mbid"
                title: "Song Title"
                length: "03:45"
                track_number: 1
                album: "Album Name"
                artist: "Artist Name"
        "400":
          description: Song ID is required
        "404":
          description: Song not found
        "500":
          description: Server error
  /api/songs:
    get:
      summary: Get list of songs (with optional filtering)
      parameters:
        - name: title
          in: query
          schema:
            type: string
        - name: artist
          in: query
          schema:
            type: string
        - name: album
          in: query
          schema:
            type: string
        - name: id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of songs
          content:
            application/json:
              example:
                - id: 11
                  mbid: "mbid-11"
                  title: "Song A"
                  album: "Album A"
                  artist: "Artist X"
                - id: 12
                  mbid: "mbid-12"
                  title: "Song B"
                  album: "Album B"
                  artist: "Artist Y"
        "400":
          description: Limit cannot exceed 500
        "500":
          description: Server error
  /api/song:
    patch:
      summary: Update a song
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: "Updated Song Title"
              length: "4:22"
              track_number: 3
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated song
          content:
            application/json:
              example:
                id: 1
                title: "Updated Song Title"
                length: "262"
                track_number: 3
        "400":
          description: No fields to update
        "401":
          description: Missing or invalid token
        "403":
          description: Insufficient permissions
        "404":
          description: Song not found
        "500":
          description: Server error

    post:
      summary: Create a new song
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: "New Song"
              album_id: 10
              length: "03:55"
              track_number: 2
              mbid: "new-mbid"
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Song created
          content:
            application/json:
              example:
                id: 50
                title: "New Song"
                album_id: 10
                length: "235"
                track_number: 2
                mbid: "new-mbid"
        "400":
          description: Missing required fields or invalid album_id
        "401":
          description: Missing or invalid token
        "403":
          description: Insufficient permissions
        "409":
          description: Song with this MBID already exists
        "500":
          description: Server error

    delete:
      summary: Delete a song
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Song deleted
          content:
            application/json:
              example:
                message: "Song deleted"
                song:
                  id: 1
                  title: "Deleted Song"
                  album_id: 10
        "401":
          description: Missing or invalid token
        "403":
          description: Insufficient permissions
        "404":
          description: Song not found
        "500":
          description: Server error
...